requirements.txt
streamlit
psycopg[binary]
python-dotenv
pandas
openpyxl
extra-streamlit-components
